renaming components for later purpose
