+++ thespoon/src/actions/modalAction.js
@@ -2,5 +2,5 @@ import * as types from "./actionTypes";
 
 export const setAuthentificatonModalVisibilityFilterAction = filter => ({
     type: types.SET_AUTHENTIFICATION_MODAL_VISIBILITY_FILTER,
-    filter
+    filter: filter
 });
