+++ thespoon/src/constants/modalVisibiltyFilters.js
