+++ thespoon/src/containers/CustomModal.js
@@ -1,10 +1,10 @@
 import React, {Component} from "react";
-import {authentificationModalVisibilityFilters} from "../constants/authentificationModalVisibiltyFilters";
+import {modalVisibilityFilters} from "../constants/modalVisibiltyFilters";
 import {roles} from "../constants/roles";
 import {connect} from "react-redux";
 import ChooseRoleModal from "../components/authentification/ChooseRoleModal";
 import {Modal} from "react-bootstrap";
-import {setAuthentificatonModalVisibilityFilterAction} from "../actions/authentificationModalAction";
+import {setModalVisibilityFilterAction} from "../actions/modalAction";
 import LogIn from "../components/authentification/LogIn";
 import RegisterRestaurantowner from "../components/authentification/RegisterRestaurantowner";
 import FillRestaurantInfo from "../components/authentification/FillRestaurantInfo"
@@ -13,12 +13,12 @@ import RegisterCustomer from "../components/authentification/RegisterCustomer";
 
 const mapStateToProps = (state) => {
     return {
-        authentificationModalVisibilityFilter: state.authentificationModalVisibilityFilter
+        modalVisibilityFilter: state.modalVisibilityFilter
     }
 }
 
 const mapDispatchToProps = dispatch => ({
-    handleClose: () => dispatch(setAuthentificatonModalVisibilityFilterAction(authentificationModalVisibilityFilters.HIDE_ALL))
+    handleClose: () => dispatch(setModalVisibilityFilterAction(modalVisibilityFilters.HIDE_ALL))
 })
 
 
@@ -26,28 +26,28 @@ class AuthentificationModal extends Component {
 
     getVisibleModal = filter => {
         switch (filter) {
-            case authentificationModalVisibilityFilters.SHOW_LOGIN:
+            case modalVisibilityFilters.SHOW_LOGIN:
                 return (
                     <LogIn onHide={() => this.props.handleClose()} />
                 )
-            case authentificationModalVisibilityFilters.SHOW_CHOOSE_ROLE:
+            case modalVisibilityFilters.SHOW_CHOOSE_ROLE:
                 return (
                     <ChooseRoleModal onHide={() => this.props.handleClose()} />
                 )
-            case authentificationModalVisibilityFilters.SHOW_REGISTER_RESTAURANT_OWNER:
+            case modalVisibilityFilters.SHOW_REGISTER_RESTAURANT_OWNER:
                 return (
                     <RegisterRestaurantowner
                         role={roles.RESTAURANT_OWNER}
                         onHide={() => this.props.handleClose()}/>
                 )
-            case authentificationModalVisibilityFilters.SHOW_RESTAURANT_INFORMATION:
+            case modalVisibilityFilters.SHOW_RESTAURANT_INFORMATION:
                 return (
                     <FillRestaurantInfo
                         role={roles.RESTAURANT_OWNER}
                         onHide={() => this.props.handleClose()}
                         />
                 )
-            case authentificationModalVisibilityFilters.SHOW_REGISTER_CUSTOMER:
+            case modalVisibilityFilters.SHOW_REGISTER_CUSTOMER:
                 return (
                     <RegisterCustomer
                         role={roles.CUSTOMER}
@@ -61,9 +61,9 @@ class AuthentificationModal extends Component {
     render() {
         return (
             <>
-                {this.props.authentificationModalVisibilityFilter !== authentificationModalVisibilityFilters.HIDE_ALL &&
+                {this.props.modalVisibilityFilter !== modalVisibilityFilters.HIDE_ALL &&
                     <Modal show={true} onHide={() => this.props.handleClose()} centered>
-                        {this.getVisibleModal(this.props.authentificationModalVisibilityFilter)}
+                        {this.getVisibleModal(this.props.modalVisibilityFilter)}
                     </Modal>
                 }
             </>
